Use Musicbrainz to fetch album info
Last.fm API is really inconsistent with giving the duration of each
track (it's often zero). The Musicbrainz ID for each track is also often
absent. The approach I'm taking now is to fetch the album info from
Musicbrainz and then do a best-match for track name to get the duration
of each track.

Thanks for being terrible, Last.fm