Finally build database containing all durations
LastFM has a terrible API.

In the end I decided to fetch each release (album) from MusicBrainz
based on the album name returned by LastFM and then do fuzzy matching on
the track names to get the correct durations.

LastFM is incredibly inconsistent with returning MusicBrainz IDs and
many of the IDs are invalid.

Paging is also weird as we get duplicates for the same timestamp
occasionally.

Special case for the commercial single release of "Photograph" as this
was a weird one to search for.

If LastFM just returned the durations itself none of this would be
necessary.

I hate you, LastFM.