Remove Rails and just build out static HTML
Now that it's all working it doesn't really make any sense at all for
this to be a Rails web app. The LastFM/Musicbrainz aggregation step is
pretty slow so it needs to be a periodic task rather than an on-demand
task and, since the statistics won't update until the aggregation has
run, it doesn't make any sense to serve this page from Rails and a
database.

Instead we just provide a Thor command to build out the HTML and then we
can push that to a static content host like Surge.sh

Now I can install this on my host and set a daily cron job to fetch all
the latest data from LastFM and push to Surge - much easier.